Add test to help understand why it acts the way it does.
The idea is to see the search-dir as 'standing on the ceiling' which is
why the ceiling isn't applied anymore allowing the search dir to escape
through the ceiling.

This is how git does it anyway, and we have to follow suit. By default,
for us this is an error though due to the configuration of
`match_ceiling_dir_or_error`.