fix: don't over-estimate the number of objects to be received in the indexing phase.
Previously, it would work with a number that is pessimistically duplicated as each object
could in theory be a ref-delta, which then causes another base to be added to the stream,
duplicating the amount of objects we need to account for in the acceleration data structure,
which unfortunatelly cannot actually grow dynamically without violating invariants.

Now we simply don't try to estimata the number of ref-deltas (which don't exist here anymore)
at this stage.