fix!: add `ReadLineBufRead::readline_str(&mut String)` to get consistent readline behaviour.
Previously the `async` version of the transport implementation would use the default `read_line` implementation
which would search for newlines in packetlines. However, these aren't always emitted which could lead to
multiple packetlines to be concatenated.

Now the we always use our implementation which simply treats one packet line as line, no matter whether
or not it ends with optional newline.