fix: it's no error if credential helpers don't receive context as input.
They might ignore all input and output hardcoded credentials for good reason,
so we don't validate writes to the helper similarly to how git does it.

Note that we also don't care if credentials are stored or erased for similar reasons.
I would happily make the case that we should be more adamant about error handling in that
case, but typically we check error codes and that's the only indicator for success or failure.

Related to https://github.com/rust-lang/cargo/issues/11821