fix!: packet-line progress loop is now interruptible.
WithSidebands may loop internally to consume remote progress, but that
doesn't play well if this reader is supposed to be interruptible, as
interrupt checks are implemented on the outer layer, not on the innermost one.

Now the progress handler can return a flag indicating whether to continue or
abort, which allows the caller to check for interrupts, and thus abort the inner
loop.