index: make `jj debug reindex` actually reindex
I broke `jj debug reindex` in a27da7d8d56d. From that commit, we no
longer delete the pointer to the old index, so nothing happens when we
reload the index. This commit fixes that, and also makes the command
error out if run on a repo with a non-default index type.