fix: uncategorized refs should not be per-worktree
When creating a new ref from using a Repository instantiated from a linked
worktree, if the ref did not fit one of the well-known categories, it would be
incorrectly stored as a per-worktree reference instead of a shared reference.

For example, the reference "refs/stacks/linked" does not fit any of the well
known categories. According to git-worktree(1), this ref should be shared
across all worktrees; i.e. located at $GIT_COMMON_DIR/refs/stacks/linked and
*not* in $GIT_COMMON_DIR/worktrees/linked/refs/stacks/linked.

This change tweaks the control flow so that uncategorized refs correctly
fallback to being relative to commondir while still placing refs that are
supposed to be per-worktree in the per-worktree location.