fix: set permissions of newly written loose objects to be similar to `git`.
Note that the current implementation lacks all of the sophistication that git
applies, and doing this properly definitely takes more work as we would need
to support `core.sharedRepository`.

Further, our tempfile implementation doesn't allow the setup of file modes
right when it matters, so that could mean quite some work to either workaround
or contribute.